Retain the instance on the view model to reduce complexity (#7329)
